ai_server/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
pillow
numpy
scikit-learn
pydantic
